+++ test/routing_test.rb
@@ -582,40 +582,6 @@ class TranslateRoutesTest < ActionController::TestCase
     assert_routing '/es/tr_parametro', controller: 'people', action: 'index', locale: 'es'
     assert_unrecognized_route '/ru/tr_param', controller: 'people', action: 'index', locale: 'ru'
   end
-
-  def test_action_controller_test_case_reads_default_urls
-    test_case_reads_default_urls(ActionController::TestCase)
-  end
-
-  def test_action_view_test_case_reads_default_urls
-    test_case_reads_default_urls(ActionView::TestCase)
-  end
-
-  def test_action_mailer_test_case_reads_default_urls
-    test_case_reads_default_urls(ActionMailer::TestCase)
-  end
-
-  private
-
-  def test_case_reads_default_urls(klass)
-    config_default_locale_settings 'en'
-
-    draw_routes do
-      localized do
-        resources :person
-      end
-    end
-
-    test_case = klass.new(nil)
-
-    # Not localized
-    assert test_case.respond_to?(:people_path)
-    assert test_case.respond_to?(:new_person_path)
-
-    # Localized
-    assert test_case.respond_to?(:people_en_path)
-    assert test_case.respond_to?(:new_person_en_path)
-  end
 end
 
 class ProductsControllerTest < ActionController::TestCase